--- test/integration/util.go
@@ -1,44 +1,0 @@
-// +build integration
-
-/*
-Copyright 2020 The Crossplane Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-	http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package integration
-
-import (
-	"context"
-	"time"
-)
-
-func waitFor(ctx context.Context, interval time.Duration, check func() (bool, error)) error {
-	ticker := time.NewTicker(interval)
-
-	for {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		case <-ticker.C:
-			// check functions return both a boolean and error to demonstrate
-			// whether an error has occurred and whether the function should
-			// continue to be called. waitFor will only exit when the check
-			// function returns true.
-			done, err := check()
-			if done {
-				return err
-			}
-		}
-	}
-}
